Created MailingListApi and added MailingListRepository, MailingListService, MailingListServiceImpl classes. Added setting to application.properties. Added new method for ApplicationConfig class. Added two dependency for pom.xml. Created two api, one for sending email to all subscribers and another is for saving subscribers to data base. Added MessageSendingException and handled it via ExceptionHandler. Lastly Added SimpleResponse.Added Validation and Added two MailingListRequest and MailingListSubscriberRequest classes. Added ModelMapper dependency

+++ src/main/java/com/example/gadgetariumb8/db/repository/MailingListRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MailingListRepository extends JpaRepository<MallingList, Integer> {
+public interface MailingListRepository extends JpaRepository<MallingList, Long> {
 }
