+++ src/main/java/com/example/gadgetariumb8/db/repository/MailingListSubscriberRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MailingListSubscriberRepository extends JpaRepository<MallingListSubscriber, Integer> {
+public interface MailingListSubscriberRepository extends JpaRepository<MallingListSubscriber, Long> {
 }
